refactor: extract shared formatting utilities to lib/format.ts
- Move calculateAge, formatGender, formatSmoking, formatAlcohol,
  formatDate, formatPhone to single module
- Remove duplicated implementations from PatientDashboard,
  SpecialistDashboard, PatientForm, register-specialist,
  update-specialist
- Replace inline phone masking with formatPhone in all forms
